# api/openapi.yaml
openapi: 3.0.3
info:
  title: 出差车票管理系统 API
  version: 1.0.0
  description: 面向百万用户出差车票管理系统的车票、审批、风控、报表和搜索接口。
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /tickets:
    get:
      summary: 查询车票列表
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
      responses:
        '200':
          description: 车票分页结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketPageResponse'
    post:
      summary: 创建车票
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /tickets/{ticketId}:
    get:
      summary: 查询车票详情
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TicketId'
      responses:
        '200':
          description: 车票详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /approvals/tickets/{ticketId}/actions:
    post:
      summary: 执行审批动作
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TicketId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalActionRequest'
      responses:
        '200':
          description: 审批后的车票
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
  /risk/events:
    get:
      summary: 查询风险事件
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: 风险事件列表
  /reports/summary:
    get:
      summary: 查询仪表盘摘要
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: 摘要指标
  /search/tickets:
    get:
      summary: 使用 Elasticsearch 搜索车票
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 200
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchPageResponse'
        '503':
          description: Elasticsearch 未启用
  /search/tickets/reindex:
    post:
      summary: 按租户重建 Elasticsearch 索引
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: 返回重建文档数量
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
    TicketId:
      name: ticketId
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    TicketStatus:
      type: string
      enum:
        - PENDING_REVIEW
        - APPROVED
        - REJECTED
        - REIMBURSED
        - EXCEPTION
        - MISSING_ATTACHMENT
    RiskLevel:
      type: string
      enum:
        - NONE
        - LOW
        - MEDIUM
        - HIGH
        - CRITICAL
    CreateTicketRequest:
      type: object
      required:
        - employeeId
        - ticketNo
        - departureCity
        - arrivalCity
        - carrierNo
        - amount
      properties:
        employeeId:
          type: integer
          format: int64
        ticketNo:
          type: string
        externalSource:
          type: string
        externalTicketId:
          type: string
        travelType:
          type: string
          example: TRAIN
        departureCity:
          type: string
          example: 上海
        arrivalCity:
          type: string
          example: 北京
        carrierNo:
          type: string
          example: G12
        seatClass:
          type: string
        departAt:
          type: string
          format: date-time
        arriveAt:
          type: string
          format: date-time
        amount:
          type: number
        currency:
          type: string
          example: CNY
    ApprovalActionRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum:
            - approve
            - reject
            - return
        comment:
          type: string
    Ticket:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tenantId:
          type: integer
          format: int64
        employeeId:
          type: integer
          format: int64
        ticketNo:
          type: string
        route:
          type: string
        departureCity:
          type: string
        arrivalCity:
          type: string
        carrierNo:
          type: string
        amount:
          type: number
        currency:
          type: string
        status:
          $ref: '#/components/schemas/TicketStatus'
        riskLevel:
          $ref: '#/components/schemas/RiskLevel'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TicketPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Ticket'
        page:
          type: integer
        size:
          type: integer
        total:
          type: integer
          format: int64
    TicketResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/Ticket'
    TicketPageResponse:
      type: object
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        data:
          $ref: '#/components/schemas/TicketPage'
    SearchPageResponse:
      allOf:
        - $ref: '#/components/schemas/TicketPageResponse'